requirements.txt
--extra-index-url https://download.pytorch.org/whl/cu117
torch==2.2.0
kornia>=0.6.9
pytorch-lightning<2.0.0
torchmetrics==0.11.0
torch-optimizer>=0.3.0
opencv-python-headless>=4.6.0
pandas>=1.3.0
numpy>=1.24.0
iterative-stratification>=0.1.7
hydra-core>=1.3.0
hydra-colorlog>=1.2.0
transformers>=4.21.3
onnx>=1.13.0
onnxruntime>=1.13.1
lightning
mlflow >= 1.20.2
albumentations>=1.3.1
